Add Footer Container and move quick add logic to composable
